+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Adds handling for multiple workspace roots
   - For colors that are named the same the tree item is expandable and shows each color with the same name
 - Better parsing of SCSS variables
+- Add the ability to name a color from within the command pallette
+  - Copy that color name to your clipboard
 - Store generated colors in a VS Code managed directory
 - Add configuration to allow copying color values instead of just variable name
 - Adds ESLint and Prettier for code formatting and linting